+++ src/main.js
@@ -1,4 +1,4 @@
-const { app, BrowserWindow, ipcMain } = require('electron');
+const { app, BrowserWindow, ipcMain, Tray, Menu } = require('electron');
 const path = require('path');
 const url = require('url');
 const packageJson = require('../package.json');
@@ -24,6 +24,7 @@ console.log(`Starting Material Companion v${packageJson.version}`);
 process.env['ELECTRON_DISABLE_SECURITY_WARNINGS'] = 'true';
 
 let win;
+let tray = null;
 
 const defaultSettings = {
     language: 'en',
@@ -31,13 +32,14 @@ const defaultSettings = {
     runInTray: false,
     mpComPort: "",
     dockComPort: "",
-    enableMpUsb: false,
     clientData: [],
     autoScanUsb: false,
     autoScanMidi: true,
     blockedMkDevices: [],
-    sensorConnectionMode: 'module',
-    mpSensorIp: 'materialsensor.local:3000'
+    sensorConnectionEvent: 'module',
+    sensorConnectionMethod: 'wifi',
+    mpSensorIp: 'materialsensor.local:3000',
+    mpSensorPreReleases: false
 }
 
 setDefaultSettings();
@@ -49,18 +51,28 @@ async function setDefaultSettings(force = false) {
     }
 }
 
-app.whenReady().then(() => {
+app.on('ready', async () => {
+    const runInTray = await settings.get('runInTray');
+    configureTray(runInTray);
     createWindow();
     setTimeout(()=> {
         initializeWebsocket(win);
         initializeMidi(win);
     }, 1000)
-    
-})
+});
 
-const createWindow = () => {
+app.on('activate', () => {
+    if (win === null) {
+        createWindow();
+    }
+});
+
+async function createWindow() {
+    const runInTray = await settings.get('runInTray');
+
     //Create window
     win = new BrowserWindow({
+        show: false,
         width: 800,
         height: 800,
         icon: path.join(__dirname, 'app', 'images', 'icons', 'png','48x48.png'),
@@ -105,6 +117,32 @@ const createWindow = () => {
             devMode: 'true'
         });
     }
+
+    win.once('ready-to-show', async () => {
+        if (!runInTray) win.show();
+    })
+
+    win.on('minimize',async function(event){
+        const runInTray = await settings.get('runInTray');
+        
+        if (runInTray) {
+            event.preventDefault();
+            win.hide();
+        }
+    });
+    
+    win.on('close', async function (event) {
+        const runInTray = await settings.get('runInTray');
+        if(runInTray && !app.isQuiting){
+            event.preventDefault();
+            win.hide();
+        }
+        return false;
+    });
+
+    win.on('closed', () => {
+        win = null;
+    })
 }
 
 function relaunchApp() {
@@ -112,7 +150,26 @@ function relaunchApp() {
     app.exit();
 }
 
+function configureTray(en) {
+    if (en) {
+        tray = new Tray(path.join(__dirname, 'app', 'images', 'icons', 'png','48x48.png'))
+        const contextMenu = Menu.buildFromTemplate([
+        { label: 'Open', click() { win.show() } },
+        { label: 'Quit', click() { app.quit() }}
+        ])
+        tray.setToolTip('Material Server')
+        tray.setContextMenu(contextMenu)
 
+        tray.on('click', () => {
+            if (!win.isVisible()) win.show();
+        })
+    }
+    else {
+        if (tray != null) tray.destroy();
+        tray = null;
+    }
+        
+}
 
 //let app.gameWindow;
 
@@ -192,7 +249,7 @@ ipcMain.handle('restart', (event) => {
 
 ipcMain.handle('setSetting', async (event, key, value) => {
     //console.log('setSetting',key,value)
-    if (key == 'sensorConnectionMode') {
+    if (key == 'sensorConnectionEvent') {
         if (value == 'module') {
             const mpModuleConnections = app.connections.filter(c => c.source == 'MaterialPlane_Foundry');
             if (mpModuleConnections.length > 0 && !app.mpClient.connected()) {