+++ changelog.md
@@ -1,4 +1,20 @@
 # Changelog Material Companion
+### v1.0.3 - 17-02-2024
+
+Additions:
+<ul>
+<li>Material Plane DIY sensors can now be updated and configured using Material Companion</li>
+<li>Material Plane hardware variants are no longer hard-coded but based on GitHub releases</li>
+<li>Added option to filter firmware and webserver pre-releases for the Material Plane sensor</li>
+<li>Added option to update the Material Plane sensor over WiFi</li>
+</ul>
+
+Fixes:
+<ul>
+<li>Removed packaged pymcuprog that would give virus warnign messages and wasn't used anymore</li>
+<li>'Run in Tray' option now works</li>
+</ul>
+
 ### v1.0.2 - 19-10-2023
 Fixes:
 <ul>